+++ src/App.js
@@ -1,15 +1,15 @@
-import { useState } from "react";
 import { InputWithLabel } from "./components/InputWithLabel/InputWithLabel";
 import { Input } from "./components/Input/Input";
 import { Checkbox } from "./components/Checkbox/Checkbox";
 import { Radio } from "./components/Radio/Radio";
 import { Button } from "./components/Button/Button";
-import { CheckboxWithLabel } from "./components/CheckboxWithLabel/CheckboxWithLabel";
-import { RadioWithLabel } from "./components/RadioWithLabel/RadioWithLabel";
+import { DropdownCheckbox } from "./components/DropdownCheckbox/DropdownCheckbox";
+import { DropdownRadio } from "./components/DropdownRadio/DropdownRadio";
+import { DropdownPageNumber } from "./components/DropdownPageNumber/DropdownPageNumber";
+import { DropdownDeleteButtons } from "./components/DropdownDeleteButtons/DropdownDeleteButtons";
+import { DropdownSelectTheme } from "./components/DropdownSelectTheme/DropdownSelectTheme";
 
 import { ReactComponent as CheckmarkIcon } from "./icons/checkmark.svg";
-import { ReactComponent as SunIcon } from "./icons/sun.svg";
-import { ReactComponent as MoonIcon } from "./icons/moon.svg";
 
 import "./style/reset.css";
 import "./style/input.css";
@@ -19,32 +19,6 @@ import "./style/button.css";
 import "./style/dropdown.css";
 
 function App() {
-  const dropdownCheckbox = [
-    { label: "Новый", checked: false },
-    { label: "Расчет", checked: false },
-    { label: "Подтвержден", checked: false },
-    { label: "Отложен", checked: false },
-    { label: "Выполнен", checked: false },
-    { label: "Отменен", checked: false },
-  ];
-
-  const [radioChoice, setRadioChoice] = useState("dropdownChoice3");
-  const [checkChoices, setCheckChoices] = useState(dropdownCheckbox);
-
-  function handleChangeRadio(e) {
-    setRadioChoice(e.target.id);
-  }
-
-  const handleChangeCheck = (index) => {
-    setCheckChoices(
-      checkChoices.map((checkbox, currentIndex) =>
-        currentIndex === index
-          ? { ...checkbox, checked: !checkbox.checked }
-          : checkbox
-      )
-    );
-  };
-
   return (
     <div className="App">
       <div class="wrapper">
@@ -133,91 +107,14 @@ function App() {
 
       <div class="wrapper__dropdown">
         <div class="wrapper__elements">
-          <div class="dropdown">
-            {checkChoices.map((checkChoice, index) => (
-              <CheckboxWithLabel
-                key={checkChoice.label}
-                checked={checkChoice.checked}
-                onChange={() => handleChangeCheck(index)}
-                label={checkChoice.label}
-                index={index}
-              />
-            ))}
-          </div>
-
-          <div class="dropdown">
-            <RadioWithLabel
-              label="Новый"
-              name="dropdownRadio"
-              id="dropdownChoice1"
-              checked={radioChoice === "dropdownChoice1"}
-              onChange={handleChangeRadio}
-            />
-            <RadioWithLabel
-              label="Расчет"
-              name="dropdownRadio"
-              id="dropdownChoice2"
-              checked={radioChoice === "dropdownChoice2"}
-              onChange={handleChangeRadio}
-            />
-            <RadioWithLabel
-              label="Подтвержден"
-              name="dropdownRadio"
-              id="dropdownChoice3"
-              checked={radioChoice === "dropdownChoice3"}
-              onChange={handleChangeRadio}
-            />
-            <RadioWithLabel
-              label="Отложен"
-              name="dropdownRadio"
-              id="dropdownChoice4"
-              checked={radioChoice === "dropdownChoice4"}
-              onChange={handleChangeRadio}
-            />
-            <RadioWithLabel
-              label="Выполнен"
-              name="dropdownRadio"
-              id="dropdownChoice5"
-              checked={radioChoice === "dropdownChoice5"}
-              onChange={handleChangeRadio}
-            />
-            <RadioWithLabel
-              label="Отменен"
-              name="dropdownRadio"
-              id="dropdownChoice6"
-              checked={radioChoice === "dropdownChoice6"}
-              onChange={handleChangeRadio}
-            />
-          </div>
+          <DropdownCheckbox />
+          <DropdownRadio />
         </div>
 
         <div class="wrapper__elements wrapper__elements_space-between">
-          <div class="dropdown dropdown_page-number">
-            <InputWithLabel
-              label="Номер страницы"
-              placeholder="Введите номер"
-            />
-          </div>
-
-          <div class="dropdown dropdown_buttons">
-            <label> Удалить n записей? </label>
-            <Button variant="blue" size="small" fullWidth>
-              Удалить
-            </Button>
-            <Button variant="white" size="small" fullWidth>
-              Отмена
-            </Button>
-          </div>
-
-          <div class="dropdown dropdown_buttons">
-            <label> Выберите тему </label>
-            <Button variant="blue" size="small" icon={SunIcon} fullWidth>
-              Светлая
-            </Button>
-            <Button variant="white" size="small" icon={MoonIcon} fullWidth>
-              Темная
-            </Button>
-          </div>
+          <DropdownPageNumber />
+          <DropdownDeleteButtons />
+          <DropdownSelectTheme />
         </div>
       </div>
     </div>
