--- src/components/Dropdown/Dropdown.js
@@ -1,8 +1,0 @@
-import cx from "classnames";
-import styles from "./Dropdown.module.css";
-
-export const Dropdown = ({ children, className, ...props }) => (
-  <div className={cx(styles._, { [className]: !!className })} {...props}>
-    {children}
-  </div>
-);
