+++ src/components/DropdownPageNumber/DropdownPageNumber.jsx
@@ -1,0 +1,11 @@
+import { InputWithLabel } from "../InputWithLabel/InputWithLabel";
+import cx from "classnames";
+import styles from "./DropdownPageNumber.module.css";
+
+export const DropdownPageNumber = ({ className, ...props }) => {
+  return (
+    <div className={cx(styles.dropdown, className)}>
+      <InputWithLabel label="Номер страницы" placeholder="Введите номер" />
+    </div>
+  );
+};
